requirements.txt
fastapi
uvicorn
python-dotenv
azure-storage-blob
azure-ai-formrecognizer
azure-search-documents
openai
python-multipart
PyJWT